replaced the home page image with url which sucks but if we can fix it up then we should
