--- client/src/home.js
@@ -1,19 +1,0 @@
-function Home() {
-    return (
-      <main>
-        <h1>Book Boulevard</h1>
-        <div >
-          <img height="300" width="500" src="client/src/books/Homepage cover 1 .jpg" alt="Book Cover" />
-          <div>
-            Photo by <a href="Thought Catalog">Cover Page</a> on <a href="https://tinyurl.com/4ssvmnnh">Unsplash</a>
-          </div>
-        </div>
-        <a href="/places">
-          <button className="btn-primary">Home Page</button>
-        </a>
-      </main>
-    );
-  }
-  
-  export default Home;
-
